backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
pydantic-settings==2.1.0
qrcode[pil]==7.4.2
python-barcode[images]==0.15.1
pillow==10.1.0

# Testing & Quality
pytest==8.0.0
pytest-asyncio==0.23.6
pytest-cov==4.1.0

# Security Scanning
bandit==1.7.7
safety==2.3.4

# Code Quality
black==24.1.0
flake8==7.0.0
mypy==1.8.0